Rename commit placement
